bpf: fix bpf_prog_info.map_ids array printing
map_id_buf had an incorrect storage size (8 bytes instead of 4), which led
to incorrect fetching of map_ids array elements, so only odd/even-indexed
elements (on big/little-endian architectures, respectively) were printed.

* src/bpf.c (print_bpf_prog_info): Change map_id_buf type from uint64_t
to uint32_t.
* tests/bpf-obj_get_info_by_fd.c: Create second bpf map, use it in the
bpf program, and add additional checks regarding map_ids printing.
* NEWS: Mention this.

Fixes: v4.23~99 "bpf: decode bpf_attr.info field used by BPF_OBJ_GET_INFO_BY_FD command"

+++ src/bpf.c
@@ -659,7 +659,7 @@ print_bpf_prog_info(struct tcb * const tcp, uint32_t bpf_fd,
 {
 	struct bpf_prog_info_struct info = { 0 };
 	const unsigned int len = MIN(size, bpf_prog_info_struct_size);
-	uint64_t map_id_buf;
+	uint32_t map_id_buf;
 
 	memcpy(&info, info_buf, len);
